+++ lib/theme_provider.dart
@@ -5,3 +5,4 @@ export 'src/provider/theme_provider.dart';
 export 'src/controller/theme_command.dart';
 
 export 'src/widgets/cycle_theme_icon_button.dart';
+export 'src/widgets/theme_dialog.dart';